services/api-gateway/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0

